Issue #7: remove total entries; move to heex files
